requirements.txt
streamlit==1.36.0
google-generativeai==0.7.1
python-dotenv==1.0.1
playwright==1.45.0 # For browser automation
python-docx==1.1.0 # For generating MS Word documents
Pillow==10.4.0 # For handling image bytes